dash/requirements.txt
snowflake-connector-python
dash
pandas
plotly
posit-sdk
